requirements.txt
openai==0.28.0
Pillow==10.0.1
Pillow==10.0.1
pytesseract==0.3.10
